move tpm submodules into the main tpm module.
why: to make the idl simpler
