--- interface/src/tpm/mod.rs
@@ -1,5 +1,0 @@
-mod tpm_dev;
-mod usr_tpm;
-
-pub use usr_tpm::*;
-pub use tpm_dev::*;
